fixed entities and docker files attacked
