+++ packages/backend/.dockerignore
@@ -1,0 +1,3 @@
+# ignore or exclude build directory
+*/dist*
+*/*/dist*
