+++ packages/backend/src/product/product.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ProductEntity } from '../entities';
+import { ProductEntity } from '@package/entities';
 import { UploadService } from '../upload/upload.service';
 import { ProductController } from './product.controller';
 import { ProductService } from './product.service';
