+++ packages/entities/src/base.entity.ts
