+++ packages/entities/src/base.interface.ts
