+++ packages/entities/src/product/index.ts
