+++ packages/entities/src/user/user.interface.ts
