+++ packages/entities/tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "extends": "../../tsconfig.json",
+    "compilerOptions": {
+        "outDir": "./dist",
+        "baseUrl": "./",
+        "paths": {
+            "@package/*": ["../*/src/index"]
+        }
+    },
+    "include": ["./src/*"],
+    "exclude": ["./dist"]
+}
