+++ packages/ui/dockerfile
@@ -1,0 +1,17 @@
+# Base image
+FROM node
+# Make folder to put our files in
+RUN mkdir -p /usr/src/app
+RUN mkdir -p /usr/src/app/ui
+# Set working directory so that all
+# subsequent command runs in this folder
+WORKDIR /usr/src/app/ui
+# Copy package json and install dependencies
+COPY package*.json ./
+RUN npm install
+# Copy our app
+COPY . .
+# Expose port to access server
+EXPOSE 6000
+# Command to run our app
+CMD [ "npm", "start" ]
